+++ main_lachtan.py
@@ -19,15 +19,13 @@ from stats_compare import get_before_stat, get_last_stat
 from main import *
 import datetime
 
-# cogs
-from bot_lachtan_spacex import spacex_launch_to_mars, rockets_shop, buy_rocket, user_wallet, create_user_acc
-# cogs
 
-all_messages = []
+# all_messages = []
 ready_list = []
 html_ready_list = []
 playing_users = []
 final_users_list = []
+session_names_ids = []
 
 # Intents
 intents = discord.Intents().all()
@@ -37,11 +35,11 @@ bot = commands.Bot(intents=intents, command_prefix='!', help_command=None)
 
 PRAYERS = 'Otče náš, ktorý si na nebesiach, posväť sa meno tvoje, príď kráľovstvo tvoje, buď vôľa tvoja ako v nebi, tak i na zemi. Chlieb náš každodenný daj nám dnes a odpusť nám naše viny, ako i my odpúšťame svojim vinníkom, a neuveď nás do pokušenia, ale zbav nás zlého.\nAmen.\n:church::cross::church:'
 
+BONUS_WEAPONS = [('total_kills_knife', 'Knife'), ('total_kills_ak47', 'AK-47'), ('total_kills_m4a1', 'M4A4'), ('total_kills_awp', 'AWP'), ('total_kills_p90', 'P90'), ('total_kills_glock', 'Glock'), ('total_kills_deagle', 'Deagle'), ('total_kills_negev', 'Negev'), ('total_kills_xm1014', 'XM-1014'), ('total_hits_mag7', 'MAG-7')]
+
+STATS_FOR_STAT_COMMAND = [('session', 'Session'), ('datetime', 'Datetime'), ('total_kills', 'Kills'), ('total_deaths', 'Deaths'), ('total_kills_headshot', 'HS kills'), ('total_damage_done', 'Demage'), ('total_kills_knife', 'Knife kills'), ('total_kills_ak47', 'AK-47 kills'), ('total_kills_m4a1', 'M4A4 kills'), ('total_kills_awp', 'AWP kills'), ('total_planted_bombs', 'Planted bombs'), ('total_defused_bombs', 'Defused bombs')]
+
 html_dict = {}
-
-kills_dict = {}
-
-list_for_i_message = {}
 
 session_dict = {}
 
@@ -49,29 +47,26 @@ POSITIVE = ["yes", "Yes", "yea", "y", "sure", "ye", "Ye", "Y"]
 MINIMAL_NUMBER = 1
 MAXIMAL_NUMBER = 11
 MAXIMAL_LEN_GAME = 30
-GAMING_ROLE = 825875949413072957
-GAMING_CATEGORY = 825872608091832332
+SESSION_ROLE = 881509999623409724 #TEST-881509999623409724, ORIG-881505468667797554
+GAMING_ROLE = 881509999623409725 #TEST-881509999623409725, ORIG-825875949413072957
+GAMING_CATEGORY = 881522734776078366 #TEST-881522734776078366, ORIG-825872608091832332
 SPACEX_ROLE = 853289230784004096
 NERD_ROLE = 853289397616640030
 
 
-@bot.command()
+@bot.command(help='All commands')
 async def help(ctx, args=None):
-    help_mes = ['Calling 911...', 'Any problem', 'Some problem here?', 'No worries, i gotch ya']
-    help_embed = discord.Embed(title=random.choice(help_mes))
+    help_mes = ['Calling 911...', 'Any problem?', 'Some problem here?', 'No worries, i gotch ya']
+    help_embed = discord.Embed(title=random.choice(help_mes), description='List of supported commands:')
     command_names_list = [x.name for x in bot.commands]
 
     if not args:
-        help_embed.add_field(
-            name="List of supported commands:",
-            value="\n> **!".join(x.name for i,x in enumerate(bot.commands))+"**",
-            inline=False
-        )
-        help_embed.add_field(
-            name="Details",
-            value="Type `!help <command name>` for more details about each command.",
-            inline=False
-        )
+        for i,x in enumerate(bot.commands):
+            help_embed.add_field(
+                name=f"`!{x.name}` {x.signature}",
+                value=f"> {x.help}",
+                inline=True
+            )
 
     # If the argument is a command, get the help text from that command:
     elif args in command_names_list:
@@ -90,62 +85,10 @@ async def help(ctx, args=None):
     await ctx.send(embed=help_embed)
 
 
-# @bot.command()
-# async def event(ctx):
-#     # if ctx.author.id != 673099696629350416:
-#     #     return
-#     await ctx.message.delete()
-#     # emojis = ['<:facts:853333168177414145>', '<:spaceX:853325340466479112>']
-#     embed = discord.Embed(title="Wassup everyone!", description="I'm sure you know I, <@783750741700509766>, am active. I've noticed that you may be interrupted by messages I send every day in #facts and #spacex. Therefore, I will leave it on you whether you want or not receive news from these two topics.", color=0xffffff)
-#     embed.add_field(name="Easy entry, Easier exit!", value="React with emojis below this message. You can remove your subscription at any time by removing your reaction.\n\n• <:facts:853333168177414145> **for facts subscription**\n• <:spaceX:853325340466479112> **for SpaceX subscription**")
-#     embed.set_footer(text='Click ⬇️ ! / Unclick ⬇️ !')
-#     # mes = await ctx.send(content="@everyone", embed=embed)
-#     # for emoji in emojis:
-#     #     await mes.add_reaction(emoji)
-#     ms = await ctx.channel.fetch_message(853339565426868234)
-#     await ms.edit(embed=embed)
-
-
-@bot.command(aliases=['balance', 'doge'])
-async def wallet(ctx):
-    await user_wallet(ctx)
-    await asyncio.sleep(5)
-    await ctx.message.delete()
-
-
-@bot.command()
-async def shop(ctx):
-    await rockets_shop(ctx)
-    await asyncio.sleep(5)
-    await ctx.message.delete()
-
-@bot.command()
-async def buy(ctx, rocket):
-    await buy_rocket(ctx, rocket)
-    await asyncio.sleep(5)
-    await ctx.message.delete()
-
-@bot.command()
-# @commands.cooldown(1, 30, commands.BucketType.user)
-async def launch(ctx, rocket):
-    await spacex_launch_to_mars(ctx, rocket)
-    await ctx.message.delete()
-
-
-@launch.error
-async def command_name_error(ctx, error):
-    if isinstance(error, commands.CommandOnCooldown):
-        remaining_time = str(datetime.timedelta(seconds=int(error.retry_after)))
-
-        em = discord.Embed(title=f"SpaceX rocket is not ready!",description=f"{ctx.author.mention}, I understand your passion for spacex but the rocket needs maintance. Try again in " + str(remaining_time), color=0xff0000)
-        await ctx.send(embed=em, delete_after=5)
-    await ctx.message.delete()
-
-
-@bot.command(help="Creates custom session with max players.", aliases=['s', 'ss'])
+@bot.command(help="Creates session", aliases=['s', 'ss'])
 async def startsession(message, name: str, max_players: int):
     if str(message.channel) != 'bot-commands':
-        await message.channel.send(f'Check the god damn channel! Nevermind, i will do it for you, its `{message.channel.name}`')
+        await message.channel.send(f'')
         return
     if max_players <= MINIMAL_NUMBER:
         await message.channel.send('Error: Minimum players in the session is **2**!')
@@ -159,16 +102,9 @@ async def startsession(message, name: str, max_players: int):
         return
     if f'{name}_game' in session_dict:
         return
-    all_messages.append(message.message)
+
     await set_vars_to_dict(message, name, max_players)
     await get_message(message, name)
-    # await move_to(session_dict[f'{name}_voice_channel'].id, message)
-
-
-@bot.command()
-async def get_server_icon_url(ctx):
-    icon_url = ctx.guild.icon_url
-    await ctx.send(f"The icon url is: {icon_url}")
 
 
 @bot.event
@@ -178,35 +114,30 @@ async def on_raw_reaction_add(payload):
     if payload.guild_id is None:
         return
 
+    flat_a = flatten(session_names_ids)
+
+    if payload.message_id not in flat_a:
+        return
+
+    for name, id in session_names_ids:
+        if payload.message_id == id:
+            session = name
+
     user = bot.get_user(payload.user_id)
     channel = await bot.fetch_channel(payload.channel_id)
-
-    if str(payload.emoji) == '<:facts:853333168177414145>':
-        if payload.channel_id == 853324149523218432:
-            await channel.guild.get_member(user.id).add_roles(channel.guild.get_role(NERD_ROLE))
-            embed = discord.Embed(title='Facts subscription!', description='Wassup nerd, I have some spicy Facts for you!', color=0x4aeb1e)
-            embed.set_footer(text='Enjoy!')
-            await user.send(embed=embed)
-
-    if str(payload.emoji) == '<:spaceX:853325340466479112>':
-        if payload.channel_id == 853324149523218432:
-            await channel.guild.get_member(user.id).add_roles(channel.guild.get_role(SPACEX_ROLE))
-            embed = discord.Embed(title='SpaceX subscription!', description='You are now following the news about SpaceX!', color=0x4aeb1e)
-            embed.set_footer(text='Enjoy!')
-            await user.send(embed=embed)
-
-    session = html_dict["session"]
     message = await channel.fetch_message(session_dict[f'{session}_first_ms_id'])
     session_owner = session_dict[f'{session}_started_by_id']
+    max_p = session_dict[f'{session}_max_players']
+    current_p = session_dict[f'{session}_ready_players']
 
-    if payload.message_id != session_dict[f'{session}_first_ms_id']:
+    if max_p == current_p:
         return
 
     if str(payload.emoji) == '❌':
         await session_unready(payload, session)
         await message.remove_reaction('❌', user)
     if str(payload.emoji) == '✅':
-        if payload.user_id in ready_list and payload.member.name != "CerveneTlacitko":
+        if payload.user_id in session_dict[f'{session}_ready_list'] and payload.member.name != "CerveneTlacitko":
             return
         session_dict[f'{session}_ready_players'] += 1
         await get_message(payload, session)
@@ -230,59 +161,27 @@ async def on_raw_reaction_add(payload):
         session_dict[f'{session}_soviet_session'] = True
 
 
-@bot.event
-async def on_raw_reaction_remove(payload):
-    if payload.user_id == bot.user.id:
-        return
-    if payload.guild_id is None:
-        return
-
-    user = bot.get_user(payload.user_id)
-    channel = await bot.fetch_channel(payload.channel_id)
-
-    if str(payload.emoji) == '<:facts:853333168177414145>':
-        if payload.channel_id == 853324149523218432:
-            await channel.guild.get_member(user.id).remove_roles(channel.guild.get_role(NERD_ROLE))
-            embed = discord.Embed(title='Unfollowed, Facts', description='I hope you learned something new!', color=0xf40101)
-            await user.send(embed=embed)
-
-    if str(payload.emoji) == '<:spaceX:853325340466479112>':
-        if payload.channel_id == 853324149523218432:
-            await channel.guild.get_member(user.id).remove_roles(channel.guild.get_role(SPACEX_ROLE))
-            embed = discord.Embed(title='Unfollowed, SpaceX', description='I hope you enjoyed SpaceX news!', color=0xf40101)
-            await user.send(embed=embed)
-
-
-@bot.command(help='delete a channel with the specified name', aliases=['e', 'end'])
+@bot.command(help='Ends session', aliases=['e', 'end'])
 async def endsession(ctx, session):
-    await session_end(ctx, session)
-
-
-async def session_end(ctx, session):
-    possible_q = ['really?', '100% sure?', 'for real?', 'are you sure?']
-    default_txt = await bot.fetch_channel(777200768842858506)
-
-    choosen_q = random.choice(possible_q)
-    channel = ctx.channel
-    dscrpt = f"Session: {session}"
-
-    embed = discord.Embed(title=f"**{ctx.author.name}**, {choosen_q}", description=dscrpt)
-    embed.set_footer(text='Reply to this message!')
-    await default_txt.send(embed=embed)
-
+    if str(ctx.channel) != 'bot-commands':
+        return
     try:
-        message = await bot.wait_for("message", check=lambda m: m.channel == channel, timeout=30.0)
-
-    except asyncio.TimeoutError:
+        session_dict[f'{session}_game']
+    except:
+        await ctx.send(f'There is no session called **"{session}"**', delete_after=5)
+        await ctx.message.delete()
+        return
+    if ctx.author.id not in session_dict[f'{session}_ready_list']:
+        embed = discord.Embed(
+            title='Missing permissions!', 
+            description='You must be in this session to end it!', 
+            color=0xff1414
+        )
+        await ctx.message.delete()
+        await ctx.send(embed=embed, delete_after=10)
         return
 
-    else:
-        if message.content in POSITIVE:
-            embed2 = discord.Embed(title='I hope you enjoyed this session', description=f'Session: **{session}**\nStatus: **ended**\nCheck <#842840247980261486> for session statistics!')
-            await message.channel.send(embed=embed2)
-            await end_function(ctx, session)
-        else:
-            return
+    await session_end(ctx, session)
 
 
 @bot.command(aliases=['m'])
@@ -290,26 +189,21 @@ async def move(ctx, session):
     await move_to_session_vc(ctx, session)
 
 
-@bot.command(aliases=["joinsession"])
+@bot.command(help='Adds one more space to session', aliases=["joinsession"])
 async def join_to(message, session: str):
     if session != session_dict[f'{session}_voice_channel']:
         return
     await ask_to_join(message, session)
 
 
-@bot.command(help='Voice channel bot join')
+@bot.command(help='Voice channel join')
 async def bot_join(ctx):
     channel = ctx.author.voice.channel
     await channel.connect()
 
 
-@bot.command()
-async def test(message, session: str):
-    get_last_stat(PLAYERS, session)
-
-
-@bot.command(help='Voice channel bot leave')
-async def leavee(ctx):
+@bot.command(help='Voice channel disconnect')
+async def leave(ctx):
     await ctx.voice_client.disconnect()
 
 
@@ -323,19 +217,19 @@ async def cheater(message):
     await command_cheater(message)
 
 
-@bot.command(help="If u forget to pray, `!pochvalen` command is for you")
+@bot.command(help="Pray, pray and pray!")
 async def pochvalen(message):
     await command_pochvalen(message)
 
 
-@bot.command(aliases=['project'])
+@bot.command(help='Detailed info about the project', aliases=['project'])
 async def info(ctx):
     toml_open = toml.load(open('pyproject.toml'))['tool']['poetry']
     project_name = toml_open['name']
     project_version = toml_open['version']
     project_description = toml_open['description']
     project_authors = ', '.join(toml_open['authors'])
-    embed = discord.Embed(title='Info about project', url='https://github.com/chalankolachtanovity/Bot_ready/blob/master/README.md')
+    embed = discord.Embed(title='Info about the project', url='https://github.com/chalankolachtanovity/Bot_ready/blob/master/README.md')
     embed.add_field(name='Name:', value=toml_open['name'])
     embed.add_field(name='Version:', value=toml_open['version'])
     embed.add_field(name='Description:', value=toml_open['description'])
@@ -343,7 +237,7 @@ async def info(ctx):
     await ctx.send(embed=embed)
 
 
-@bot.command()
+@bot.command(help='Share your music')
 async def spotify(ctx, user: discord.Member = None):
     if user is None:
         user = ctx.author
@@ -363,34 +257,48 @@ async def spotify(ctx, user: discord.Member = None):
             await ctx.send(embed=embed)
 
 
-async def create_vc_channel(guild, game: str, number: int):
-    allowed_role = discord.utils.get(guild.roles, name="gaming")
-    banned_role = discord.utils.get(guild.roles, name="MUNI")
-    banned_role_1 = discord.utils.get(guild.roles, name="lachtan")
-    overwrites = {
-        allowed_role: discord.PermissionOverwrite(connect=True),
-        banned_role: discord.PermissionOverwrite(connect=False),
-        banned_role_1: discord.PermissionOverwrite(connect=False),
+@bot.command(help='See your last session stats')
+async def stats(ctx, user: discord.Member = None):
+    if user is None:
+        user = ctx.author
+        pass
+    name = get_username(user.name)
 
+    embed = discord.Embed(title=f'Last session stats for **{name}**')
+    for sql_name, nice_name in STATS_FOR_STAT_COMMAND:
+        embed.add_field(name=f'> {nice_name}', value=f'> {get_stat(sql_name, name)}')
+    await ctx.send(embed=embed)
+
+
+async def create_vc_channel(guild, game: str, number: int):
+    banned_role = discord.utils.get(guild.roles, name="@everyone")
+
+    overwrites = {
+        banned_role: discord.PermissionOverwrite(connect=False),
     }
+    for user_id in session_dict[f'{game}_ready_list']:
+        user = await bot.fetch_user(user_id)
+        overwrites[user] = discord.PermissionOverwrite(connect=True)
+
     new_voice_channel = await guild.create_voice_channel(name=f"{game} session", category=bot.get_channel(GAMING_CATEGORY), user_limit=number+1, overwrites=overwrites)
     session_dict[f'{game}_voice_channel'] = new_voice_channel
     if session_dict[f'{game}_soviet_session'] is True:
-        await russian_anthem(new_voice_channel)
+        await soviet_songs(new_voice_channel)
 
 
-async def russian_anthem(channel):
-    paths_to_anthem = ['mp3/soviet_anthem_0.mp3', 'mp3/soviet_anthem_1.mp3']
+async def soviet_songs(channel):
+    paths_to_anthem = ['mp3/soviet_anthem.mp3', 'mp3/Konarmejskij-marsh.mp3', 'mp3/Nash-paravos.mp3', 'mp3/Varshavjanka.mp3']
     vc = await channel.connect()
     vc.play(discord.FFmpegPCMAudio(random.choice(paths_to_anthem)), after=lambda e: print('done', e))
-    await asyncio.sleep(195)
+    await asyncio.sleep(120)
     await vc.disconnect()
+    await channel.edit(user_limit=channel.user_limit-1)
 
 
 async def set_vars_to_dict(message, game: str, number: int):
-    list_for_i_message[f"{game}_ready_list"] = []
     html_dict["session"] = game
     html_dict["max_players"] = number
+    session_dict[f"{game}_ready_list_mentions"] = []
     session_dict[f'{game}_started_by_id'] = message.author.id
     session_dict[f'{game}_playing_users'] = []
     session_dict[f'{game}_ready_list'] = []
@@ -420,23 +328,31 @@ async def get_message(message, session):
 
 async def first_session_message(message, started_session: str):
     """Sends starting session message"""
-    welcome_to_session = ['This is', 'Say hi to', 'Wanna join?', "I'ts gonna be fun with ", '"Here comes the"', 'Something is waiting for you to join -', "Very tempting, isn't it?", "Now, you can't stop it", 'you, are, addicted...']
 
-    if session_dict[f'{started_session}_gaming_session'] is True:
-        await message.guild.get_member(message.author.id).add_roles(message.guild.get_role(GAMING_ROLE))
-        await game_presence(started_session)
+    await in_lobby_presence(started_session)
 
     max_players = session_dict[f'{started_session}_max_players']
 
-    embed = discord.Embed(title=f"**{random.choice(welcome_to_session)} {started_session} session**", description=f"Started by **{message.author}**\nSession: **{started_session}**\nPlayers able to join: **{max_players}**")
+    embed = discord.Embed(
+        title=f"**{started_session}** session - lobby", 
+        description=f"""
+        Started by **{message.author}**
+        Session: **{started_session}**
+        Players able to join: **{max_players}**
+        Players joined: {message.author.mention}
+        """,
+        color=0x1f8b4c
+    )
     embed.set_footer(text=f"Join with emoji reaction below ↓")
     first_session_ms = await message.channel.send(embed=embed)
     session_dict[f'{started_session}_first_ms_id'] = first_session_ms.id
+
     emojis = ['❌', '✅', '📊', '🔊', '🎮', '<:ouremoji:851164594320048208>']
     for emoji in emojis:
         await first_session_ms.add_reaction(emoji)
 
-    list_for_i_message[f"{started_session}_ready_list"].append(message.author.mention)
+    session_names_ids.append((started_session, first_session_ms.id))
+    session_dict[f"{started_session}_ready_list_mentions"].append(message.author.mention)
     ready_list.append(message.author.id)
     html_ready_list.append(message.author.name)
 
@@ -447,109 +363,190 @@ async def session_ready(payload, session):
         return
     guild = bot.get_guild(payload.guild_id)
     member = guild.get_member(payload.user_id)
-
-    # session_dict[f'{session}_ready_players'] += 1
-    # await move_to(session_dict[f'{session_to}_voice_channel'].id, message)
+    channel = await bot.fetch_channel(payload.channel_id)
+    started_by = await bot.fetch_user(session_dict[f'{session}_started_by_id'])
 
     session_dict[f'{session}_ready_list'].append(member.id)
-    list_for_i_message[f"{session}_ready_list"].append(member.mention)
+    session_dict[f"{session}_ready_list_mentions"].append(member.mention)
+    member_mentions = ', '.join(session_dict[f'{session}_ready_list_mentions'])
 
-    if session_dict[f'{session}_gaming_session'] is True:
-        await guild.get_member(member.id).add_roles(guild.get_role(GAMING_ROLE))
+    msg = await channel.fetch_message(session_dict[f'{session}_first_ms_id'])
+
+    embed = discord.Embed(
+        title=f"**{session}** session - lobby", 
+        description=f"""
+        Started by **{started_by}**
+        Session: **{session}**
+        Players able to join: **{session_dict[f'{session}_max_players']}**
+        Players joined: {member_mentions}
+        """,
+        color=0x1f8b4c
+    )
+    embed.set_footer(text=f"Join with emoji reaction below ↓")
+    await msg.edit(embed=embed)
 
     ready_list.append(member.id)
     html_ready_list.append(member.name)
 
-@bot.command()
-async def ooo(ctx):
-  await ctx.channel.send('⏱️')
-  await ctx.message.add_reaction('⏱️')
-  # add timer option -> on_reaction_event... assign_fucker_role
-  # test -> it worked...
-
 
 async def ending_of_session(payload, ended_session: str):
     """Sends last ready information message, execute timer to assign fucker and delete all things that was in before session"""
-    ending_messages = ["Let's go!", 'Full of upcoming fun!', 'Players - registred.', 'All ready', 'Prepare for battle', "Players, don't pee your pants", "Let's save the world", "Let's blow up the world"]
 
     guild = bot.get_guild(payload.guild_id)
     member = guild.get_member(payload.user_id)
     channel = await bot.fetch_channel(payload.channel_id)
+    msg = await channel.fetch_message(session_dict[f'{ended_session}_first_ms_id'])
 
     session_dict[f'{ended_session}_ready_list'].append(member.id)
-    list_for_i_message[f"{ended_session}_ready_list"].append(member.mention)
-
-    after_ready_list_members = ', '.join(list_for_i_message[f'{ended_session}_ready_list'])
+    session_dict[f"{ended_session}_ready_list_mentions"].append(member.mention)
+    after_ready_list_members = ', '.join(session_dict[f'{ended_session}_ready_list_mentions'])
     all_p = session_dict[f'{ended_session}_max_players']
-    embed = discord.Embed(title=f'**{ended_session}** session. {random.choice(ending_messages)}', description=f'Players: **{all_p}**/**{all_p}**\nPlayers list: {after_ready_list_members}')
-    embed.set_footer(text="Enjoy this session! Don't forget to launch CS:GO!")
-    await channel.send(embed=embed)
 
-    if session_dict[f'{ended_session}_create_vc'] is True:
-        number = session_dict[f'{ended_session}_max_players']
-        await create_vc_channel(guild, ended_session, number)
+    embed = discord.Embed(
+        title=f'{ended_session} session - Full', 
+        description=f"""
+        Status: **Ongoing**
+        Players: **{all_p}**/**{all_p}**
+        Players list: {after_ready_list_members}
+        """,
+        color=0x9E0000#0x9b59b6
+        )
+    embed.set_footer(text="Enjoy this session! Don't forget to launch CS:GO!")
+
+    if session_dict[f'{ended_session}_gaming_session'] is True:
+        for member_id in session_dict[f'{ended_session}_ready_list']:
+            await guild.get_member(member_id).add_roles(guild.get_role(GAMING_ROLE))
+        embed.set_thumbnail(url="https://www.dictionary.com/e/wp-content/uploads/2018/07/CS-GO3.jpg")
+
+    if session_dict[f'{ended_session}_soviet_session'] is True:
+        embed.set_thumbnail(url="https://i.pinimg.com/originals/f4/67/a2/f467a22305243ef1eee7373ca8e0f0b9.png")
+
+    await msg.edit(embed=embed)
+    await session_presence(ended_session)
 
     if session_dict[f'{ended_session}_statistics'] is True:
         for name in html_ready_list:
             check_list(name)
         get_before_stat(final_users_list)
 
-    for single_message in all_messages:
-        await single_message.delete()
+        session_dict[f'{ended_session}_bonus_weapon'] = [random.choice(BONUS_WEAPONS)]
+        for env, nice in session_dict[f'{ended_session}_bonus_weapon']:
+            embed = discord.Embed(title=f'Bonus weapon for {ended_session} session is {nice}  🎁', description='The player with most kills on this weapon will win this session')
+        await channel.send(embed=embed)
 
-    if session_dict[f'{ended_session}_gaming_session'] is True:
-        for member_id in session_dict[f'{ended_session}_ready_list']:
-            await guild.get_member(member_id).add_roles(guild.get_role(GAMING_ROLE))
-    await default_presence()
+    if session_dict[f'{ended_session}_create_vc'] is True:
+        number = session_dict[f'{ended_session}_max_players']
+        await create_vc_channel(guild, ended_session, number)
 
-    html_dict.pop("session")
-    html_dict.pop("max_players")
-    all_messages.clear()
+    for member_id in session_dict[f'{ended_session}_ready_list']:
+        await guild.get_member(member_id).add_roles(guild.get_role(SESSION_ROLE))
+
+    html_dict.clear()
     ready_list.clear()
     html_ready_list.clear()
 
 
 def check_list(user):
     if user == 'CerveneTlacitko':
-        final_users_list.append(('t_kmaso', 'Kmasko', KMASKO_DICT['personaname']))
+        final_users_list.append(('t_kmaso', 'Kmasko', KMASKO_DICT['personaname'], 673099696629350416))
     if user == 'teetou':
-        final_users_list.append(('t_teetou', 'Teetou', TEETOU_DICT['personaname']))
+        final_users_list.append(('t_teetou', 'Teetou', TEETOU_DICT['personaname'], 765908540878487572))
     if user == 'Bonsai':
-        final_users_list.append(('t_aligator', 'Aligator', ALIGATOR_DICT['personaname']))
+        final_users_list.append(('t_aligator', 'Aligator', ALIGATOR_DICT['personaname'], 489114476222742528))
     if user == 'milan čonka':
-        final_users_list.append(('t_stano', 'Stano', STANO_DICT['personaname']))
+        final_users_list.append(('t_stano', 'Stano', STANO_DICT['personaname'], 489117343964725260))
     if user == 'Tajmoti':
-        final_users_list.append(('t_tajmoti', 'Tajmoti', TAJMOTI_DICT['personaname']))
+        final_users_list.append(('t_tajmoti', 'Tajmoti', TAJMOTI_DICT['personaname'], 140197100062244864))
     if user == 'andrzej':
-        final_users_list.append(('t_dron', 'Dron', DRON_DICT['personaname']))
+        final_users_list.append(('t_dron', 'Dron', DRON_DICT['personaname'], 491236539716730890))
+    if user == 'Kulivox':
+        final_users_list.append(('t_kulivox', 'Kulivox', KULIVOX_DICT['personaname'], 265552128335020032))
+
+
+def get_username(username):
+    usernames = [('CerveneTlacitko', 'Kmasko'), ('teetou', 'Teetou'), ('Bonsai', 'Aligator'), ('milan čonka', 'Stano'), ('Tajmoti', 'Tajmoti'), ('andrzej', 'Dron'), ('Kulivox', 'Kulivox')]
+    for d_name, sql_name in usernames:
+        if d_name == username:
+            return sql_name
+
+
+async def session_end(ctx, session):
+    possible_q = ['really?', '100% sure?', 'for real?', 'are you sure?']
+    emojis = ['✅', '❌']
+
+    choosen_q = random.choice(possible_q)
+    dscrpt = f"Session: {session}"
+
+    embed = discord.Embed(title=f"**{ctx.author.name}**, {choosen_q}", description=dscrpt)
+    embed.set_footer(text='Add reaction below ↓')
+    confirming_message = await ctx.send(embed=embed)
+    for emoji in emojis:
+        await confirming_message.add_reaction(emoji)
+
+    def check(reaction, user):
+        return user == ctx.author
+
+    try:
+        reaction, user = await bot.wait_for('reaction_add', timeout=30.0, check=check)
+
+    except asyncio.TimeoutError:
+        await ctx.message.delete()
+        await confirming_message.delete()
+        return
+
+    else:
+        if str(reaction.emoji) == '✅':
+            await confirming_message.delete()
+            await ctx.message.delete()
+            await end_function(ctx, session)
+        else:
+            await ctx.message.delete()
+            await confirming_message.delete()
 
 
 async def end_function(ctx, vc_name):
     existing_channel = discord.utils.get(ctx.guild.channels, name=f"{vc_name} session")
     general_channel = discord.utils.get(ctx.guild.channels, name="General")
-    if existing_channel is None:
-        await ctx.send(f'No channel named "{vc_name}" was found')
-        return
+    msg = await ctx.fetch_message(session_dict[f'{vc_name}_first_ms_id'])
+
+    if session_dict[f'{vc_name}_create_vc'] == True:
+        for member in ctx.author.voice.channel.members:
+            await member.move_to(general_channel)
+            if existing_channel is None:
+                await ctx.send(f'No channel named "{vc_name}" was found')
+                return
+
+    embed = discord.Embed(
+        title=f'{vc_name} session - Ended', 
+        description=f"""
+        Status: **Ended**
+        Session: **{vc_name}**
+        Check <#842840247980261486> for session statistics!
+        """,
+    )
+    await msg.edit(embed=embed)
+
     for member in session_dict[f'{vc_name}_ready_list']:
         user = bot.get_user(member)
         await ctx.guild.get_member(user.id).remove_roles(ctx.guild.get_role(GAMING_ROLE))
-
-    for member in ctx.author.voice.channel.members:
-        await member.move_to(general_channel)
+        await ctx.guild.get_member(user.id).remove_roles(ctx.guild.get_role(SESSION_ROLE))
 
     if existing_channel is not None:
         await existing_channel.delete()
     if session_dict[f'{vc_name}_statistics'] is True:
         get_last_stat(final_users_list, vc_name)
-    await asyncio.sleep(5)
+        await asyncio.sleep(5)
     if session_dict[f'{vc_name}_statistics'] is True:
-        await create_stats_table()
+        await create_stats_table(vc_name)
+    ms_id = session_dict[f'{vc_name}_first_ms_id']
+    session_names_ids.remove((vc_name, ms_id))
     await pop_items_in_dict(vc_name)
+    await default_presence()
     final_users_list.clear()
 
 
 async def pop_items_in_dict(session):
-    keys_to_pop = [f'{session}_started_by_id', f'{session}_playing_users', f'{session}_ready_list', f'{session}_game', f'{session}_max_players', f'{session}_ready_players', f'{session}_statistics', f'{session}_create_vc', f'{session}_gaming_session', f'{session}_voice_channel', f'{session}_first_ms_id', f'{session}_soviet_session']
+    keys_to_pop = [f'{session}_started_by_id', f'{session}_playing_users', f'{session}_ready_list', f'{session}_game', f'{session}_max_players', f'{session}_ready_players', f'{session}_statistics', f'{session}_create_vc', f'{session}_gaming_session', f'{session}_voice_channel', f'{session}_first_ms_id', f'{session}_soviet_session', f'{session}_bonus_weapon', f'{session}_ready_list_mentions']
     for key in keys_to_pop:
         session_dict.pop(key, None)
 
@@ -569,17 +566,45 @@ async def ask_to_join(message, session: str):
     channel = session_dict[f'{session}_voice_channel']
     embed = discord.Embed(title=f"{message.author.name}, you can join", description=f'Session: {session}')
     await message.channel.send(embed=embed)
-    await channel.edit(user_limit=channel.user_limit + 1)
-    await message.guild.get_member(message.author.id).add_roles(message.guild.get_role(GAMING_ROLE))
+    user = await bot.fetch_user(message.author.id)
+    overwrites = {
+        user: discord.PermissionOverwrite(connect=True)
+    }
+    await channel.edit(user_limit=channel.user_limit+1, overwrites=overwrites)
+
+    await message.guild.get_member(message.author.id).add_roles(message.guild.get_role(SESSION_ROLE))
+    if session_dict[f'{session}_gaming_session'] is True:
+        await message.guild.get_member(message.author.id).add_roles(message.guild.get_role(GAMING_ROLE))
 
 
-async def create_stats_table():
-    channel = await bot.fetch_channel(842840247980261486)
-    myEmbed = discord.Embed(title='Statistics for **{}** session'.format(get_stat('session', 'Kmasko')))
-    for env, nice, steam in final_users_list:
-        myEmbed.add_field(name=f'**{steam}**', value='> Kills: {}\n> Deaths: {}'.format(get_stat('total_kills', (nice)), get_stat('total_deaths', (nice))), inline=True)
-        myEmbed.set_footer(text=get_stat('datetime', 'Kmasko'))
-    await channel.send(embed=myEmbed)
+async def create_stats_table(session):
+    channel = await bot.fetch_channel(881509999807971359) #842840247980261486
+    embed = discord.Embed(title='Statistics for **{}** session'.format(get_stat('session', 'Kmasko')))
+    for_best_stat = {}
+    b_w = session_dict[f'{session}_bonus_weapon']
+    for env, nice in b_w:
+        bonus_weapon = nice
+        bonus_weapon_env = env
+    for env, nice, steam, user_id in final_users_list:
+        bonus_stat = get_stat(bonus_weapon_env, (nice))
+        for_best_stat[user_id] = bonus_stat
+        embed.add_field(
+            name=f'**{steam}**', 
+            value='> Kills: {}\n> Deaths: {}\n> {} kills: **{}**'.format(get_stat('total_kills', (nice)), get_stat('total_deaths', (nice)), bonus_weapon, bonus_stat), inline=True
+        )
+        embed.set_footer(text=get_stat('datetime', 'Kmasko'))
+
+    all_values = for_best_stat.items()
+    max_value = max(for_best_stat.values())
+    user_id = max(all_values, key = lambda x:x[1])
+    user = bot.get_user(user_id[0])
+
+    embed.add_field(
+        name=f'**Session winner**  🎉', 
+        value=f'> {user.mention} is the winner with **{max_value}** **{bonus_weapon}** kills', 
+        inline=False
+    )
+    await channel.send(embed=embed)
 
 
 def get_stat(stat, name):
@@ -596,7 +621,7 @@ def get_stat(stat, name):
 async def session_unready(payload, session: str):
     if session != session_dict[f'{session}_game']:
         return
-    if payload.user_id not in ready_list:
+    if payload.user_id not in session_dict[f'{session}_ready_list']:
         return
     if session_dict[f'{session}_ready_players'] == 1:
         return
@@ -604,50 +629,72 @@ async def session_unready(payload, session: str):
         playing_users.remove(payload.user_id)
     session_dict[f'{session}_ready_players'] -= 1
 
+    mentions = session_dict[f'{session}_ready_list_mentions']
     guild = bot.get_guild(payload.guild_id)
     member = guild.get_member(payload.user_id)
+    channel = await bot.fetch_channel(payload.channel_id)
+    msg = await channel.fetch_message(session_dict[f'{session}_first_ms_id'])
+    started_by = await bot.fetch_user(session_dict[f'{session}_started_by_id'])
 
+    if member.mention in mentions:
+        session_dict[f'{session}_ready_list_mentions'].remove(member.mention)
+    session_dict[f'{session}_ready_list'].remove(payload.user_id)
     ready_list.remove(payload.user_id)
     html_ready_list.remove(member.name)
+
+    player_mentions = ', '.join(session_dict[f'{session}_ready_list_mentions'])
+
+    embed = discord.Embed(
+        title=f"**{session}** session - lobby", 
+        description=f"""
+        Started by **{started_by}**
+        Session: **{session}**
+        Players able to join: **{session_dict[f'{session}_max_players']}**
+        Players joined: {player_mentions}
+        """,
+        color=0x1f8b4c
+    )
+    embed.set_footer(text=f"Join with emoji reaction below ↓")
+    await msg.edit(embed=embed)
 
 
 async def timer_assign_fucker(choose_timer: int):
     await asyncio.sleep(choose_timer)
 
 
-async def assing_fucker_role(message, game):
-    """After 5minutes it compare playing users with ready users and according to result,
-  add role to ready player but not playing"""
-    await timer_assign_fucker(120)
-    role = message.guild.get_role(812862836710834207)
-    suka_list = []
-    checker = []
-    missing_players = []
-    print("Timer ended")
-    for i in session_dict[f'{game}_ready_list']:
-        if i in playing_users:
-            checker.append(True)
-        else:
-            checker.append(i)
-    if checker.count(True) != len(checker):
-        missing_players.append(checker)
-        for f_player in missing_players[0]:
-            user = bot.get_user(f_player)
-            await message.guild.get_member(user.id).add_roles(role)
-            suka_list.append(user.mention)
-        embed = discord.Embed(title='Wall of shame!', description=f'Shame on you {suka_list[0]}!')
-        embed.set_footer(text="Get rid of this role by launching CS:GO under 2 minutes!")
-        await message.channel.send(embed=embed)
+# async def assing_fucker_role(message, game):
+#     """After 5minutes it compare playing users with ready users and according to result,
+#   add role to ready player but not playing"""
+#     await timer_assign_fucker(120)
+#     role = message.guild.get_role(812862836710834207)
+#     suka_list = []
+#     checker = []
+#     missing_players = []
+#     print("Timer ended")
+#     for i in session_dict[f'{game}_ready_list']:
+#         if i in playing_users:
+#             checker.append(True)
+#         else:
+#             checker.append(i)
+#     if checker.count(True) != len(checker):
+#         missing_players.append(checker)
+#         for f_player in missing_players[0]:
+#             user = bot.get_user(f_player)
+#             await message.guild.get_member(user.id).add_roles(role)
+#             suka_list.append(user.mention)
+#         embed = discord.Embed(title='Wall of shame!', description=f'Shame on you {suka_list[0]}!')
+#         embed.set_footer(text="Get rid of this role by launching CS:GO under 2 minutes!")
+#         await message.channel.send(embed=embed)
 
-    else:
-        print("equal")
+#     else:
+#         print("equal")
 
-    x = [value for value in ready_list if value in playing_users]
-    x_list = []
-    x_list.append(x)
-    for u_player in x_list[0]:
-        unban_user = bot.get_user(u_player)
-        await message.guild.get_member(unban_user.id).remove_roles(role)
+#     x = [value for value in session_dict[f'{game}_ready_list'] if value in playing_users]
+#     x_list = []
+#     x_list.append(x)
+#     for u_player in x_list[0]:
+#         unban_user = bot.get_user(u_player)
+#         await message.guild.get_member(unban_user.id).remove_roles(role)
 
 
 async def command_russia(message):
@@ -685,6 +732,22 @@ async def get_dir_videos(dirname: str) -> list:
     return videos
 
 
+def flatten(l, ltypes=(list, tuple)):
+    ltype = type(l)
+    l = list(l)
+    i = 0
+    while i < len(l):
+        while isinstance(l[i], ltypes):
+            if not l[i]:
+                l.pop(i)
+                i -= 1
+                break
+            else:
+                l[i:i + 1] = l[i]
+        i += 1
+    return ltype(l)
+
+
 async def command_pochvalen(message):
     """Replies 'aj naveky, amen' to the player saying pochvalen."""
     myEmbed = discord.Embed(title="Aj naveky, Amen.",
@@ -702,60 +765,64 @@ async def default_presence():
         type=discord.ActivityType.watching, name='Spaceflights'))
 
 
-async def game_presence(game):
+async def in_lobby_presence(game):
     await bot.change_presence(activity=discord.Activity(
         type=discord.ActivityType.playing, name=f'In {game} lobby'))
 
 
+async def session_presence(session):
+    await bot.change_presence(activity=discord.Activity(
+        type=discord.ActivityType.watching, name=f'{session} session'))
+
 async def send_report_message(message):
     report_channel = bot.get_channel(777200768842858506)
-    await report_channel.send(f'{message.author.mention} tried to send "** {message.content} **" in **{message.channel}** channel\nStop! Thats not channel for that stuff!')
+    await report_channel.send(f'{message.author.mention} tried to send "** {message.content} **" in **{message.channel}** channel\nYou can\'t do that')
 
 
 # REDDIT
-reddit = asyncpraw.Reddit(
-                    client_id=os.getenv('client_id'),
-                    client_secret=os.getenv('client_secret'),
-                    username=os.getenv('username'),
-                    password=os.getenv('password'),
-                    user_agent="pythonpraw"
-                    )
+# reddit = asyncpraw.Reddit(
+#                     client_id=os.getenv('client_id'),
+#                     client_secret=os.getenv('client_secret'),
+#                     username=os.getenv('username'),
+#                     password=os.getenv('password'),
+#                     user_agent="pythonpraw"
+#                     )
 
 
-async def spacex_reddit():
-    spacex_channel = bot.get_channel(818166852064247818)
-    facts_channel = bot.get_channel(819169067747246120)
-    spacex_subreddit = await reddit.subreddit("spacex")
-    facts_subreddit = await reddit.subreddit("facts")
-    async for spacex_submission in spacex_subreddit.stream.submissions(skip_existing=True):
-        extension = spacex_submission.url[len(spacex_submission.url) - 3 :].lower()
-        if "jpg" not in extension and "png" not in extension:
-            await reddit_message_img(spacex_channel, spacex_submission.subreddit,
-            spacex_submission.title, '<3', 'https://i.imgur.com/hzrRwO3.png', 1
-            )
-        else:
-            await reddit_message_img(spacex_channel, spacex_submission.subreddit, spacex_submission.title, '',spacex_submission.url, 1
-            )
+# async def spacex_reddit():
+#     spacex_channel = bot.get_channel(818166852064247818)
+#     facts_channel = bot.get_channel(819169067747246120)
+#     spacex_subreddit = await reddit.subreddit("spacex")
+#     facts_subreddit = await reddit.subreddit("facts")
+#     async for spacex_submission in spacex_subreddit.stream.submissions(skip_existing=True):
+#         extension = spacex_submission.url[len(spacex_submission.url) - 3 :].lower()
+#         if "jpg" not in extension and "png" not in extension:
+#             await reddit_message_img(spacex_channel, spacex_submission.subreddit,
+#             spacex_submission.title, '<3', 'https://i.imgur.com/hzrRwO3.png', 1
+#             )
+#         else:
+#             await reddit_message_img(spacex_channel, spacex_submission.subreddit, spacex_submission.title, '',spacex_submission.url, 1
+#             )
 
-        async for facts_submission in facts_subreddit.stream.submissions(skip_existing=True):
-            number = 3
-            await reddit_message_img(facts_channel, facts_submission.subreddit,
-            facts_submission.title, facts_submission.author, number
-            )
+#         async for facts_submission in facts_subreddit.stream.submissions(skip_existing=True):
+#             number = 3
+#             await reddit_message_img(facts_channel, facts_submission.subreddit,
+#             facts_submission.title, facts_submission.author, number
+#             )
 
 
-async def reddit_message_img(reddit_channel, subreddit_name, post_title, post_link, pos):
-    if pos == 3:
-      embed=discord.Embed(title=f"r/{subreddit_name}", description=f"{post_title}", color=0x0)
-      await reddit_channel.send(embed=embed)
-    if pos !=3:
-      if pos == 1:
-        embed=discord.Embed(title=f"r/{subreddit_name}", url=f"{post_link}", description=f"{post_title}", color=0x0)
-        await reddit_channel.send(embed=embed)
-      if pos == 2:
-        embed=discord.Embed(title=f"r/{subreddit_name}", url=f"{post_link}", description=f"{post_title}", color=0x0)
-        embed.set_image(url=f"{post_link}")
-        await reddit_channel.send(embed=embed)
+# async def reddit_message_img(reddit_channel, subreddit_name, post_title, post_link, pos):
+#     if pos == 3:
+#       embed=discord.Embed(title=f"r/{subreddit_name}", description=f"{post_title}", color=0x0)
+#       await reddit_channel.send(embed=embed)
+#     if pos !=3:
+#       if pos == 1:
+#         embed=discord.Embed(title=f"r/{subreddit_name}", url=f"{post_link}", description=f"{post_title}", color=0x0)
+#         await reddit_channel.send(embed=embed)
+#       if pos == 2:
+#         embed=discord.Embed(title=f"r/{subreddit_name}", url=f"{post_link}", description=f"{post_title}", color=0x0)
+#         embed.set_image(url=f"{post_link}")
+#         await reddit_channel.send(embed=embed)
 # REDDIT
 
 async def read_image(message):
@@ -779,10 +846,10 @@ async def on_message(message):
     if (message.author == bot.user):
         return
 
-    await create_user_acc(message)
+    # await create_user_acc(message)
 
-    # if str(message.channel) == 'test' and message.attachments[0].url is not None:
-    #     await read_image(message)
+    if str(message.channel) == 'pytess' and message.attachments[0].url is not None:
+         await read_image(message)
 
     if str(message.channel) != 'bot-commands' and message.content[0] == '!':
         if message.author.name != 'CerveneTlacitko':
@@ -837,11 +904,11 @@ async def on_member_join(member):
 #     await vc.disconnect()
 
 
-@bot.command()
-async def playy(ctx):
+@bot.command(help='Rickroll your mates')
+async def play(ctx):
     voicechannel = discord.utils.get(ctx.guild.channels, name='General') # ctx.author.voice.channel
     vc = await voicechannel.connect()
-    vc.play(discord.FFmpegPCMAudio("rickrollaf.mp3"), after=lambda e: print('done', e))
+    vc.play(discord.FFmpegPCMAudio("mp3/rickrollaf.mp3"), after=lambda e: print('rickrolled', e))
 
 
 @bot.event
@@ -860,8 +927,8 @@ async def on_command_error(ctx, error):
 async def on_ready():
     print("Discord -> connected")
     await default_presence()
-    await spacex_reddit()
-    await bot.process_commands
+    # await spacex_reddit()
+    # await bot.process_commands
 
     
 keep_alive()